chore: temporarily disable noble tests for role and playbook ibridges
Tests on Ubuntu 20 were failing for the iBridges playbook and role because the default Python on Focal is no longer supported by iBridges.
Tried to add Noble to the build instead, but the Noble build is currently failing because of bugs with the SRC-External component.
Temporarily disable the Noble tests for iBridges and test only on Jammy for now.